Flytt ferdigstillOppgaveOgKnyttJournalpostTilBehandling endepunkt til OppgaveController
